Generate Artificial data under the GUILDS model, conditioned on Guild size

Using this function it is possible to generate a community dataset consisting of two separate abundance vectors for each guild, where the data generated adhere to the Guilds model. Data generated is conditioned on guild size.

generate.Guilds.Cond(theta, alpha_x, alpha_y, JX, JY)

Arguments

theta

Fundamental Biodiversity Number theta

alpha_x

Dispersal Ability of Guild X

alpha_y

Dispersal Ability of Guild Y

JX

Total number of individuals in Guild X

JY

Total number of individuals in Guild Y

Value

guildX

Vector containing the unlabeled abundances of species in Guild X

guildY

Vector containing the unlabeled abundances of species in Guild Y

Author

Thijs Janzen

Examples

generate.Guilds.Cond(theta = 200,
                     alpha_x = 0.005,
                     alpha_y = 0.001,
                     JX = 15000,
                     JY = 5000);
#> $guildX
#>   [1] 862 637 579 551 485 462 434 422 422 413 393 369 367 344 309 290 283 282
#>  [19] 272 269 243 220 218 211 207 204 198 179 172 169 166 164 159 159 157 147
#>  [37] 142 139 137 129 120 116 113 112 108  97  87  86  81  80  78  71  66  64
#>  [55]  62  60  57  56  50  49  49  48  46  45  43  43  43  40  39  38  35  34
#>  [73]  34  32  32  32  30  29  28  25  25  24  24  23  22  21  19  17  17  16
#>  [91]  16  16  16  16  15  15  14  13  13  13  12  12  12  12  11  11  10  10
#> [109]  10   9   9   9   8   8   8   7   7   7   7   7   7   7   7   6   6   6
#> [127]   6   6   5   5   5   5   4   4   3   3   3   3   3   3   3   3   3   3
#> [145]   2   2   2   2   2   2   1   1   1   1   1   1   1   1   1   1   1   1
#> [163]   1   1   1
#> 
#> $guildY
#>  [1] 1162  652  427  283  255  251  245  180  162  129  112  103  102   93   86
#> [16]   75   75   71   53   50   37   35   34   30   27   27   23   22   18   17
#> [31]   16   15   14   14   12   12   10   10    9    9    7    5    4    4    3
#> [46]    2    2    2    2    2    1    1    1    1    1    1    1    1    1    1
#>