Maximization of the loglikelihood given the standard Neutral Model, using the Etienne Sampling Formula

This function computes the maximum likelihood estimates of the parameters of the Neutral model, using the Etienne Sampling Formula

Usage

maxLikelihood.ESF(init_vals, abund, verbose = FALSE)

Arguments

init_vals

A vector of initial starting values, of the format c(theta, m)

abund

Vector containing a record of the number of individuals per species

verbose

TRUE/FALSE flag, indicates whether intermediate output is shown on screen

Value

the output is a list containing the following:

par

a vector containing the parameter values at the maximum likelihood c(theta, m)

fvalues

the likelihood at the corresponding parameter values

conv

gives a message on convergence of optimization; conv = 0 means convergence

References

Etienne, R.S. (2005). A new sampling formula for neutral biodiversity. Ecology Letters, 8(3), 253-260.

Author

Thijs Janzen

Examples

  A <- c(1, 1, 1, 3, 5, 8)
  maxLikelihood.ESF( c(7, 0.1), abund = A)
#> $par
#> [1] 15.5921293  0.1688449
#> 
#> $value
#> [1] 5.270295
#>